added resize of modal-body based on footer and header size
